Update script.js
modified the random list generator so there is a random mask list generator and a random vaccine list generator. The random vaccine list generator ensures that patient zero does not get chosen. The only time patient zero will have a vaccine (and not get infected) will be when there is a 100% vaccine level. I also changed patient zero to only be infected if the simulation.vaccLevel is less than 100.